+++ server/src/bruFiles/diagnostics/shared/diagnosticCodes/relevantWithinAuthBlockDiagnosticCodeEnum.ts
@@ -2,8 +2,11 @@ export enum RelevantWithinAuthBlockDiagnosticCode {
     KeysMissingInAuthBlock = "bru300",
     UnknownKeysDefinedInAuthBlock = "bru301",
     DuplicateKeysDefinedInAuthBlock = "bru302",
+
     InvalidGrantType = "bru303",
+
     InvalidApiKeyAuthValueForPlacement = "bru304",
+
     InvalidOAuth2ValueForCredentialsPlacement = "bru305",
     InvalidOAuth2ValueForTokenPlacement = "bru306",
     InvalidOAuth2ValueForAutoFetchToken = "bru3078",
@@ -11,4 +14,10 @@ export enum RelevantWithinAuthBlockDiagnosticCode {
     InvalidOAuth2ValueForPkce = "bru309",
     InvalidOAuth2ValueForTokenSource = "bru310",
     AdditionalParamsFieldsDefinedWithoutOAuth2AuthType = "bru311",
+
+    AkamaiEdgeGrid_MaxBodySizeValueInvalid = "bru320",
+
+    InvalidOAuth1SignatureMethodValue = "bru330",
+    InvalidOAuth1PlacementValue = "bru331",
+    InvalidOAuth1IncludeBodyHashValue = "bru332",
 }