+++ shared/languageUtils/commonBlocks/generic/AuthTypes.ts
@@ -1,4 +1,4 @@
-export const MethodBlockAuthValues = {
+export const AuthTypes = {
     None: "none",
     Inherit: "inherit",
     Basic: "basic",
@@ -6,7 +6,9 @@ export const MethodBlockAuthValues = {
     Awsv4: "awsv4",
     Digest: "digest",
     Htlm: "ntlm",
+    Oauth1: "oauth1",
     Oauth2: "oauth2",
     Wsse: "wsse",
     Apikey: "apikey",
+    AkamaiEdgeGridAuth: "akamai-edgegrid",
 } as const;