+++ shared/languageUtils/requestFiles/requestFileBlockNameEnum.ts
@@ -25,9 +25,11 @@ export enum RequestFileBlockName {
     AwsSigV4Auth = "auth:awsv4",
     DigestAuth = "auth:digest",
     NtlmAuth = "auth:ntlm",
+    OAuth1Auth = "auth:oauth1",
     OAuth2Auth = "auth:oauth2",
     WsseAuth = "auth:wsse",
     ApiKeyAuth = "auth:apikey",
+    AkamaiEdgeGridAuth = "auth:akamai-edgegrid",
 
     // OAuth2 additional params
     OAuth2AdditionalParams_AuthReq_Headers = "auth:oauth2:additional_params:auth_req:headers",